remove post install scripts

+++ heasoft/build.sh
@@ -59,9 +59,11 @@ export HEADAS=$PREFIX/$HEA_SUBDIR
 hmake install-heasoftpy
 cd ../../BUILD_DIR
 
-# for xspec local models
-cp ../Xspec/BUILD_DIR/hmakerc $PREFIX/$HEA_SUBDIR/bin/
-cp ../Xspec/BUILD_DIR/Makefile-std $PREFIX/$HEA_SUBDIR/bin/
+# for xspec local models; check in case xspec was not requested
+if [ -f ../Xspec/BUILD_DIR/hmakerc ]; then
+    cp ../Xspec/BUILD_DIR/hmakerc $PREFIX/$HEA_SUBDIR/bin/
+    cp ../Xspec/BUILD_DIR/Makefile-std $PREFIX/$HEA_SUBDIR/bin/
+fi
 
 # Copy fix-x11-conda.sh; we are inside BUILD_DIR
 cp fix-x11-conda.sh $PREFIX/$HEA_SUBDIR/BUILD_DIR/
@@ -97,17 +99,8 @@ source \$HEADAS/BUILD_DIR/headas-init.csh
 EOF
 chmod +x $PREFIX/bin/heainit.csh
 
-cat <<EOF >$PREFIX/bin/.heasoft-post-link.sh
-mkdir -p \$CONDA_PREFIX/etc/conda/activate.d
-cp \$CONDA_PREFIX/bin/heainit.*sh \$CONDA_PREFIX/etc/conda/activate.d/
+mkdir -p $PREFIX/etc/conda/activate.d
+cp $PREFIX/bin/heainit.*sh $PREFIX/etc/conda/activate.d/
 
-mkdir -p \$CONDA_PREFIX/etc/conda/deactivate.d
-cp \$CONDA_PREFIX/$HEA_SUBDIR/BUILD_DIR/headas-uninit.*sh \$CONDA_PREFIX/etc/conda/deactivate.d/
-
-# fix conda x11 for mac
-bash \$CONDA_PREFIX/$HEA_SUBDIR/BUILD_DIR/fix-x11-conda.sh \$CONDA_PREFIX
-EOF
-cat <<EOF >$PREFIX/bin/.heasoft-pre-unlink.sh
-rm \$CONDA_PREFIX/etc/conda/activate.d/heainit.*sh > /dev/null 2>&1
-rm \$CONDA_PREFIX/etc/conda/deactivate.d/headas-uninit.*sh > /dev/null 2>&1
-EOF
+mkdir -p $PREFIX/etc/conda/deactivate.d
+cp $PREFIX/$HEA_SUBDIR/BUILD_DIR/headas-uninit.*sh $PREFIX/etc/conda/deactivate.d/
